+++ src/main/java/net/minecraft/src/BlockSoulSand.java
@@ -12,7 +12,7 @@ public class BlockSoulSand extends Block {
 	 */
 	public AxisAlignedBB getCollisionBoundingBoxFromPool(World par1World, int par2, int par3, int par4) {
 		float var5 = 0.125F;
-		return AxisAlignedBB.getAABBPool().getAABB((double) par2, (double) par3, (double) par4, (double) (par2 + 1), (double) ((float) (par3 + 1) - var5), (double) (par4 + 1));
+		return AxisAlignedBB.getAABBPool().getAABB(par2, par3, par4, par2 + 1, (float) (par3 + 1) - var5, par4 + 1);
 	}
 
 	/**
