+++ src/main/java/net/minecraft/src/EntitySquid.java
@@ -132,7 +132,7 @@ public class EntitySquid extends EntityWaterMob {
 			this.renderYawOffset += (-((float) Math.atan2(this.motionX, this.motionZ)) * 180.0F / (float) Math.PI - this.renderYawOffset) * 0.1F;
 			this.rotationYaw = this.renderYawOffset;
 			this.squidYaw += (float) Math.PI * this.field_70871_bB * 1.5F;
-			this.squidPitch += (-((float) Math.atan2((double) var1, this.motionY)) * 180.0F / (float) Math.PI - this.squidPitch) * 0.1F;
+			this.squidPitch += (-((float) Math.atan2(var1, this.motionY)) * 180.0F / (float) Math.PI - this.squidPitch) * 0.1F;
 		} else {
 			this.tentacleAngle = MathHelper.abs(MathHelper.sin(this.field_70867_h)) * (float) Math.PI * 0.25F;
 			this.squidPitch = (float) ((double) this.squidPitch + (double) (-90.0F - this.squidPitch) * 0.02D);
