+++ src/main/java/net/minecraft/src/EntityTNTPrimed.java
@@ -20,9 +20,9 @@ public class EntityTNTPrimed extends Entity {
 		this.setWorld(par1World);
 		this.setPosition(par2, par4, par6);
 		float var9 = (float) (Math.random() * Math.PI * 2.0D);
-		this.motionX = (double) (-((float) Math.sin((double) var9)) * 0.02F);
+		this.motionX = -((float) Math.sin(var9)) * 0.02F;
 		this.motionY = 0.20000000298023224D;
-		this.motionZ = (double) (-((float) Math.cos((double) var9)) * 0.02F);
+		this.motionZ = -((float) Math.cos(var9)) * 0.02F;
 		this.fuse = 80;
 		this.prevPosX = par2;
 		this.prevPosY = par4;
